shopify-middleware/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyjwt